[Feat] 나의 누적 평가 요약 조회 / 평가된 스터디 목록 조회 / 평가 상세 조회 API 구현

+++ src/main/java/navik/domain/evaluation/dto/EvaluationSubmitDTO.java
@@ -14,7 +14,7 @@ public record EvaluationSubmitDTO(
 	@NotNull
 	@DecimalMin(value = "1.0")
 	@DecimalMax(value = "5.0")
-	Float score,
+	Double score,
 
 	@Size(min = 5, max = 5, message = "이런 점이 뛰어나요 태그 5개를 선택해야 합니다.")
 	List<Long> strengthTagIds,
